Move torch functions around.
