+++ parlai/agents/bert_ranker/bi_encoder_ranker.py
@@ -5,7 +5,7 @@
 # LICENSE file in the root directory of this source tree.
 from parlai.utils.distributed import is_distributed
 from parlai.core.torch_ranker_agent import TorchRankerAgent
-from parlai.utils.misc import padded_3d
+from parlai.utils.torch import padded_3d
 from parlai.zoo.bert.build import download
 
 from .bert_dictionary import BertDictionaryAgent
