+++ parlai/agents/legacy_agents/seq2seq/modules_v1.py
@@ -15,7 +15,7 @@ from torch.nn.parameter import Parameter
 from torch.nn.utils.rnn import pad_packed_sequence, pack_padded_sequence
 import torch.nn.functional as F
 
-from parlai.utils.misc import NEAR_INF
+from parlai.utils.torch import NEAR_INF
 
 
 def opt_to_kwargs(opt):
