+++ parlai/agents/seq2seq/modules.py
@@ -15,7 +15,7 @@ from torch.nn.parameter import Parameter
 from torch.nn.utils.rnn import pad_packed_sequence, pack_padded_sequence
 import torch.nn.functional as F
 
-from parlai.utils.misc import NEAR_INF
+from parlai.utils.torch import NEAR_INF
 from parlai.core.torch_generator_agent import TorchGeneratorModel
