+++ parlai/agents/transformer/modules.py
@@ -24,7 +24,7 @@ import numpy as np
 
 from parlai.core.torch_generator_agent import TorchGeneratorModel
 from parlai.utils.misc import warn_once
-from parlai.utils.misc import neginf
+from parlai.utils.torch import neginf
 
 try:
     from apex.normalization.fused_layer_norm import FusedLayerNorm as LayerNorm
