+++ parlai/utils/misc.py
@@ -15,22 +15,17 @@ import pickle
 import random
 import time
 import traceback
-from typing import Union, Optional, Set, Tuple, Any, Dict, List, Sized
+from typing import Union, Optional, Set, Any, Dict, List
 import warnings
 
 from parlai.core.message import Message
 
-# some of the utility methods are helpful for Torch
 try:
     import torch
-    import torch.optim
 except ImportError:
-    raise ImportError('Parlai requires pytorch. Go to http://pytorch.org to install.')
-
-
-"""Near infinity, useful as a large penalty for scoring when inf is bad."""
-NEAR_INF = 1e20
-NEAR_INF_FP16 = 65504
+    # silence the error, we'll have other problems later.
+    # we only need the import for typing here
+    pass
 
 
 DISPLAY_MESSAGE_DEFAULT_FIELDS = {
@@ -48,16 +43,6 @@ DISPLAY_MESSAGE_DEFAULT_FIELDS = {
     'label_candidates_vecs',
     'token_losses',
 }
-
-
-def neginf(dtype: torch.dtype) -> float:
-    """
-    Return a representable finite number near -inf for a dtype.
-    """
-    if dtype is torch.float16:
-        return -NEAR_INF_FP16
-    else:
-        return -NEAR_INF
 
 
 def maintain_dialog_history(
@@ -427,6 +412,26 @@ class AttrDict(dict):
         self.__dict__ = self
 
 
+class NoLock(object):
+    """
+    Empty `lock`.
+
+    Does nothing when you enter or exit.
+    """
+
+    def __enter__(self):
+        """
+        No-op.
+        """
+        return self
+
+    def __exit__(self, exc_type, exc_value, exc_traceback):
+        """
+        No-op.
+        """
+        pass
+
+
 def round_sigfigs(x: Union[float, torch.Tensor], sigfigs=4) -> float:
     """
     Round value to specified significant figures.
@@ -448,26 +453,6 @@ def round_sigfigs(x: Union[float, torch.Tensor], sigfigs=4) -> float:
             return x
         else:
             raise ex
-
-
-class NoLock(object):
-    """
-    Empty `lock`.
-
-    Does nothing when you enter or exit.
-    """
-
-    def __enter__(self):
-        """
-        No-op.
-        """
-        return self
-
-    def __exit__(self, exc_type, exc_value, exc_traceback):
-        """
-        No-op.
-        """
-        pass
 
 
 single_nolock = NoLock()
@@ -909,153 +894,6 @@ def set_namedtuple_defaults(namedtuple, default=None):
     return namedtuple
 
 
-def padded_tensor(
-    items: List[Union[List[int], torch.LongTensor]],
-    pad_idx: int = 0,
-    use_cuda: bool = False,
-    left_padded: bool = False,
-    max_len: Optional[int] = None,
-    fp16friendly: bool = False,
-) -> Tuple[torch.LongTensor, List[int]]:
-    """
-    Create a right-padded matrix from an uneven list of lists.
-
-    Returns (padded, lengths), where padded is the padded matrix, and lengths
-    is a list containing the lengths of each row.
-
-    Matrix is right-padded (filled to the right) by default, but can be
-    left padded if the flag is set to True.
-
-    Matrix can also be placed on cuda automatically.
-
-    :param list[iter[int]] items: List of items
-    :param bool sort: If True, orders by the length
-    :param int pad_idx: the value to use for padding
-    :param bool use_cuda: if true, places `padded` on GPU
-    :param bool left_padded:
-    :param int max_len: if None, the max length is the maximum item length
-    :param bool fp16friendly: if True, pads the time dimension to be a multiple of 8.
-
-    :returns: (padded, lengths) tuple
-    :rtype: (Tensor[int64], list[int])
-    """
-
-    # number of items
-    n = len(items)
-    # length of each item
-    lens: List[int] = [len(item) for item in items]  # type: ignore
-    # max in time dimension
-    t = max(lens) if max_len is None else max_len
-
-    # if input tensors are empty, we should expand to nulls
-    t = max(t, 1)
-
-    if fp16friendly and (t % 8 != 0):
-        # pad to be a multiple of 8 to ensure we use the tensor cores
-        t += 8 - (t % 8)
-
-    if isinstance(items[0], torch.Tensor):
-        # keep type of input tensors, they may already be cuda ones
-        output = items[0].new(n, t)  # type: ignore
-    else:
-        output = torch.LongTensor(n, t)  # type: ignore
-    output.fill_(pad_idx)
-
-    for i, (item, length) in enumerate(zip(items, lens)):
-        if length == 0:
-            # skip empty items
-            continue
-        if not isinstance(item, torch.Tensor):
-            # put non-tensors into a tensor
-            item = torch.LongTensor(item)  # type: ignore
-        if left_padded:
-            # place at end
-            output[i, t - length :] = item
-        else:
-            # place at beginning
-            output[i, :length] = item
-
-    if use_cuda:
-        output = output.cuda()
-    return output, lens
-
-
-def padded_3d(
-    tensors: List[torch.LongTensor],
-    pad_idx: int = 0,
-    use_cuda: bool = False,
-    dtype: Optional[torch.dtype] = torch.long,
-    fp16friendly: bool = False,
-):
-    """
-    Make 3D padded tensor for list of lists of 1D tensors or lists.
-
-    :param tensors:
-        list of lists of 1D tensors (or lists)
-    :param pad_idx:
-        padding to fill tensor with
-    :param use_cuda:
-        whether to call cuda() before returning
-    :param bool fp16friendly:
-        if True, pads the final dimension to be a multiple of 8.
-
-    :returns:
-        3D tensor with the maximum dimensions of the inputs
-    """
-    a = len(tensors)
-    b = max(len(row) for row in tensors)  # type: ignore
-    c = max(len(item) for row in tensors for item in row)  # type: ignore
-
-    # pad empty tensors
-    if fp16friendly and c % 8 != 0:
-        c += 8 - (c % 8)
-    c = max(c, 1)
-
-    output = torch.full((a, b, c), pad_idx, dtype=dtype)
-
-    for i, row in enumerate(tensors):
-        item: Sized
-        for j, item in enumerate(row):  # type: ignore
-            if len(item) == 0:
-                continue
-            if not isinstance(item, torch.Tensor):
-                item = torch.Tensor(item, dtype=dtype)
-            output[i, j, : len(item)] = item
-
-    if use_cuda:
-        output = output.cuda()
-
-    return output
-
-
-def argsort(keys: List[Any], *lists: List[List[Any]], descending: bool = False):
-    """
-    Reorder each list in lists by the (descending) sorted order of keys.
-
-    :param iter keys:
-        Keys to order by.
-    :param list[list] lists:
-        Lists to reordered by keys's order.  Correctly handles lists and 1-D
-        tensors.
-    :param bool descending:
-        Use descending order if true.
-
-    :returns:
-        The reordered items.
-    """
-    ind_sorted = sorted(range(len(keys)), key=lambda k: keys[k])
-    if descending:
-        ind_sorted = list(reversed(ind_sorted))
-    output = []
-    for lst in lists:
-        # watch out in case we don't have torch installed
-        if isinstance(lst, torch.Tensor):
-            output.append(lst[ind_sorted])
-        else:
-            output.append([lst[i] for i in ind_sorted])
-    return output
-
-
 _seen_warnings: Set[str] = set()
 
 
@@ -1070,48 +908,3 @@ def warn_once(msg: str, warningtype=None) -> None:
     if msg not in _seen_warnings:
         _seen_warnings.add(msg)
         warnings.warn(msg, warningtype, stacklevel=2)
-
-
-def fp16_optimizer_wrapper(
-    optimizer: torch.optim.Optimizer,  # type: ignore
-    verbose: bool = False,
-    dynamic_loss_scale: bool = True,
-    loss_initial_scale: float = 2.0 ** 17,
-):
-    """
-    Wrap the an optimizer with FP16 loss scaling protection.
-
-    Requires apex to be installed. Will throw an ImportError if it is not.
-
-    :param optimizer:
-        Any torch optimizer
-    :param bool verbose:
-        Enables verbose output in the FP16 optimizer. Turning this on can help
-        debug when FP16 is underperforming.
-    :param bool dynamic_loss_scaling:
-        FP16 requires loss scaling to avoid underflows. It is recommended this
-        stays on, but advanced users may want it off.
-    :param float loss_initial_scale:
-        Initial loss scaling. Default chosen empirically, but models with very low
-        or high loss values may need this adjusted. Stick with powers of 2.
-
-    :returns:
-        An APEX FP16 optimizer. Please note this has different requirements on
-        how backward() and step() are called.
-    """
-    try:
-        import apex.fp16_utils
-    except ImportError:
-        raise ImportError(
-            'No fp16 support without apex. Please install it from '
-            'https://github.com/NVIDIA/apex'
-        )
-    return apex.fp16_utils.FP16_Optimizer(
-        optimizer,
-        dynamic_loss_scale=dynamic_loss_scale,
-        verbose=verbose,
-        # TODO: We may later want to remove this flag. Right now it
-        # empirically improves the first few backward passes, but future APEX
-        # improvements may make this unnecessary.
-        dynamic_loss_args={'init_scale': loss_initial_scale},
-    )
