+++ projects/controllable_dialogue/controllable_seq2seq/modules.py
@@ -15,7 +15,7 @@ from torch.nn.parameter import Parameter
 from torch.nn.utils.rnn import pad_packed_sequence, pack_padded_sequence
 import torch.nn.functional as F
 
-from parlai.utils.misc import NEAR_INF
+from parlai.utils.torch import NEAR_INF
 
 
 def opt_to_kwargs(opt):
