+++ projects/self_feeding/self_feeding_agent.py
@@ -20,7 +20,8 @@ import torch.nn as nn
 import torch.nn.functional as F
 
 from parlai.core.torch_agent import TorchAgent, Output
-from parlai.utils.misc import round_sigfigs, warn_once, padded_tensor
+from parlai.utils.misc import round_sigfigs, warn_once
+from parlai.utils.torch import padded_tensor
 
 from parlai.agents.transformer.transformer import TransformerRankerAgent
