+++ projects/wizard_of_wikipedia/generator/agents.py
@@ -13,7 +13,8 @@ from functools import lru_cache
 import torch as th
 import numpy as np
 
-from parlai.utils.misc import padded_tensor, round_sigfigs
+from parlai.utils.torch import padded_tensor
+from parlai.utils.misc import round_sigfigs
 
 from parlai.agents.transformer.transformer import TransformerGeneratorAgent
